Reference

Contents

Reference

Contents

Index

Model Components

Index

UrbanTethysChloris.ModelComponents.Parameters.BuildingEnergyModelParameters

UrbanTethysChloris.ModelComponents.Parameters.HVACParameters

UrbanTethysChloris.ModelComponents.Parameters.HeightDependentVegetationParameters

UrbanTethysChloris.ModelComponents.Parameters.IndoorOpticalProperties

UrbanTethysChloris.ModelComponents.Parameters.LocationSpecificThermalProperties

UrbanTethysChloris.ModelComponents.Parameters.OpticalProperties

UrbanTethysChloris.ModelComponents.Parameters.ParameterSet

UrbanTethysChloris.ModelComponents.Parameters.PersonParameters

UrbanTethysChloris.ModelComponents.Parameters.SimpleOpticalProperties

UrbanTethysChloris.ModelComponents.Parameters.SoilParameters

UrbanTethysChloris.ModelComponents.Parameters.SurfaceFractions

UrbanTethysChloris.ModelComponents.Parameters.ThermalBuilding

UrbanTethysChloris.ModelComponents.Parameters.ThermalProperties

UrbanTethysChloris.ModelComponents.Parameters.UrbanGeometryParameters

UrbanTethysChloris.ModelComponents.Parameters.VegetatedOpticalProperties

UrbanTethysChloris.ModelComponents.Parameters.VegetatedSoilParameters

UrbanTethysChloris.ModelComponents.Parameters.VegetationParameters

UrbanTethysChloris.ModelComponents.Parameters.WindowParameters

TethysChlorisCore.validate_fields

UrbanTethysChloris.check_extraneous_fields

TethysChlorisCore.validate_fields — Method

TethysChlorisCore.validate_fields(::Type{T}, data::Dict{String,Any}) where {T<:AbstractModelComponent}

Validate that a dictionary contains only the required fields for a given model component type.

Arguments

T: Type of model component to validate fields against

data: Dictionary containing fields to validate

Throws

ArgumentError: If any extraneous fields are found in the data dictionary

UrbanTethysChloris.check_extraneous_fields — Method

check_extraneous_fields(::Type{T}, data::Dict{String,Any}) where {T<:AbstractModelComponent}

Check if the input dictionary contains any keys that are not part of the required fields for type T.

Arguments

T: Type of model component to check fields against

data: Dictionary containing fields to validate

required_fields: List of required fields for the model component. This is typically obtained from get_required_fields(T) or fieldnames(T)

Throws

ArgumentError: If any extraneous keys are found in the data dictionary

Model Components

Parameters

UrbanTethysChloris.ModelComponents.Parameters.HeightDependentVegetationParameters — Type

HeightDependentVegetationParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the location-specific (tree, roof, ground) vegetation.

Fields

LAI::FT: Leaf area index [-]

SAI::FT: Stem area index [-].

hc::FT: Canopy height [m].

h_disp::FT: [-].

d_leaf::FT: [-].

CASE_ROOT::Int: Type of root profile [-].

ZR95::FT: Root depth, 95th percentile [mm].

ZR50::FT: Root depth, 50th percentile [mm].

ZRmax::FT: Root depth, maximum [mm]

Rrootl::FT: Root length index [m root m-2 PFT].

PsiL50::FT: Water potential at 50 % of leaf hydraulic conductivity [MPa].

PsiX50::FT: Water potential at 50 % of xylem hydraulic conductivity and limit for water extraction from soil [MPa].

FI::FT: Intrinsec quantum efficency [µmol CO2 µmol-1 photons].

Do::FT: Empirical coefficient that expresses the value of vapor pressure deficit at which f(∆e) = 0.5 [Pa].

a1::FT: Empirical parameter linking net assimilaton AnCto stomatal conductanceg{s,CO2}` [-].

go::FT: Minimum/cuticular stomatal conductance [mol CO2 m-2 leaf s-1].

CT::Int: Photosyntesis Typology for Plants, Photosynthetic pathway C3 or C4 [-]

DSE::FT: Activation Energy - Plant Dependent, Activation Energy in Photosynthesis for Rubisco Capacity [kJ/mol].

Ha::FT: Entropy factor - Plant Dependent, Activation energy [kJ / mol K].

gmes::FT: Mesophyll conductance, not used [mol CO2/s m2].

rjv::FT: Scaling factor between Jmax and V_{c,max} [µmol equivalent µmol-1 CO2].

Kopt::FT: Optical depth of direct beam per unit plant area (?) [-].

Knit::FT: Canopy nitrogen decay coefficient [-].

Vmax::FT: Maximum Rubisco capacity at 25◦C leaf scale [µmol CO2 m-2 s-1]

mSl::FT: [-].

e_rel::FT: Relative Efficiency of the photosynthesis apparatus due to Age/Day-length [-].

e_relN::FT: Relative efficiency of the photosynthesis apparatus due to N limitations [-].

Psi_sto_00::FT: Soil water potential at the beginning of stomatal closure [MPa].

Psi_sto_50::FT: Soil water potential at 50 % stomatal closure [MPa].

Sl::FT: Specific leaf area of biomass [m^2 /gC] [-].

SPARTREE::Int: Tree root distribution: 1 = Tree roots can access all water in the soil

(imp, bare, veg) equally; 2 = If the tree crown is smaller than the combined vegetated and bare fraction, then the trees only transpire from these fractions. Otherwise, they also transpire from the impervious ground fraction. [-].

UrbanTethysChloris.ModelComponents.Parameters.SurfaceFractions — Type

SurfaceFractions{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the SurfaceFractions model component.

Fields

roof::LocationSpecificSurfaceFractions{FT}: Roof surface fractions.

ground::LocationSpecificSurfaceFractions{FT}: Ground surface fractions.

UrbanTethysChloris.ModelComponents.Parameters.LocationSpecificThermalProperties — Type

LocationSpecificThermalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the location-specific (wall, roof, ground) thermal properties.

Fields

lan_dry::FT: Thermal conductivity dry solid [W/m K]

cv_s::FT: Volumetric heat capacity solid [J/m^3 K].

UrbanTethysChloris.ModelComponents.Parameters.UrbanGeometryParameters — Type

UrbanGeometryParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the UrbanGeometry model component.

Fields

Height_canyon::FT: Height of urban canyon [m].

Width_canyon::FT: Ground width of urban canyon [m].

Width_roof::FT: Roof width of urban canyon [m].

Height_tree::FT: Tree height [m].

Radius_tree::FT: Tree radius (=1/4 fg,tree *Wcan) [m].

Distance_tree::FT: Distance of wall to tree trunk [m].

Hcan_max::FT: Maximum height of roughness elements (buildings) [m].

Hcan_std::FT: Standard deviation of roughness elements (buildings) [m].

trees::Bool: Easy switch to include (=1) and exclude (=0) trees in the urban canyon.

ftree::FT: Tree fraction along canyon axis

hcanyon::FT: Normalized height of urban canyon [-].

wcanyon::FT: Normalized ground width of urban canyon [-].

wroof::FT: Normalized roof width of urban canyon [-]

htree::FT: Normalized tree height [-]

radius_tree::FT: Normalized tree radius [-].

distance_tree::FT: Normalized distance of wall to tree trunk [-].

ratio::FT: Height-to-width ratio [-].

wcanyon_norm::FT: Normalized canyon width overall [-].

wroof_norm::FT: Normalized roof width overall [-].

UrbanTethysChloris.ModelComponents.Parameters.VegetationParameters — Type

VegetationParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Container for vegetation parameters for different urban surface components.

Fields

roof::HeightDependentVegetationParameters{FT}: Vegetation parameters for roof vegetation

ground::HeightDependentVegetationParameters{FT}: Vegetation parameters for ground-level vegetation

tree::HeightDependentVegetationParameters{FT}: Vegetation parameters for trees

UrbanTethysChloris.ModelComponents.Parameters.ThermalProperties — Type

ThermalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Container for vegetation parameters for different urban surface components.

Fields

roof::LocationSpecificThermalProperties{FT}: Roof thermal properties

ground::LocationSpecificThermalProperties{FT}: Ground thermal properties

wall::LocationSpecificThermalProperties{FT}: Wall thermal properties

UrbanTethysChloris.ModelComponents.Parameters.OpticalProperties — Type

OpticalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Container for optical properties for different urban surface components.

Fields

roof::VegetatedOpticalProperties{FT}: Roof optical properties

ground::VegetatedOpticalProperties{FT}: Ground optical properties

wall::SimpleOpticalProperties{FT}: Wall optical properties

tree::SimpleOpticalProperties{FT}: Tree optical properties

UrbanTethysChloris.ModelComponents.Parameters.SimpleOpticalProperties — Type

SimpleOpticalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the location-specific (wall and tree) optical properties, for which albedo and emissivity have been pre-calculated.

Fields

albedo::FT: Surface albedo (-)

emissivity::FT: Surface emissivity (-)

UrbanTethysChloris.ModelComponents.Parameters.VegetatedOpticalProperties — Type

VegetatedOpticalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Optical properties specific to vegetated surfaces (roof and ground)

Fields

aveg::FT: Vegetation surface albedo (-)

aimp::FT: Impervious surface albedo (-)

abare::FT: Bare surface albedo (-), only used for ground

eveg::FT: Vegetation surface emissivity (-)

eimp::FT: Impervious surface emissivity (-)

ebare::FT: Bare surface emissivity (-), only used for ground

UrbanTethysChloris.ModelComponents.Parameters.BuildingEnergyModelParameters — Type

BuildingEnergyModelParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Container for all building energy model parameters.

Fields

indoor_optical::IndoorOpticalProperties{FT}: Indoor surface optical properties

thermal::ThermalBuilding{FT}: Building thermal properties

windows::WindowParameters{FT}: Window parameters

hvac::HVACParameters{FT}: HVAC system parameters

UrbanTethysChloris.ModelComponents.Parameters.IndoorOpticalProperties — Type

IndoorOpticalProperties{FT<:AbstractFloat} <: AbstractParameters{FT}

Optical properties for indoor building surfaces.

Fields

abc::FT: Albedo ceiling (-)

abw::FT: Albedo wall (-)

abg::FT: Albedo ground (-)

abm::FT: Albedo internal mass (-)

ec::FT: Emissivity ceiling (-)

eg::FT: Emissivity ground (-)

ew::FT: Emissivity wall (-)

em::FT: Emissivity internal mass (-)

UrbanTethysChloris.ModelComponents.Parameters.ThermalBuilding — Type

ThermalBuilding{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters specifying thermal properties and dimensions of building internal mass.

Fields

IntMassOn::Int: Include building internal mass in calculation (0/1)

FloorHeight::FT: Average floor height in building (m)

dzFloor::FT: Average thickness of floors in building (m)

dzWall::FT: Average thickness of walls in building (m)

lan_ground_floor::FT: Building ground thermal conductivity (W/m K)

cv_ground_floor::FT: Building ground volumetric heat capacity (J/m³ K)

lan_floor_IntMass::FT: Internal mass floor thermal conductivity (W/m K)

cv_floor_IntMass::FT: Internal mass floor volumetric heat capacity (J/m³ K)

lan_wall_IntMass::FT: Internal mass wall thermal conductivity (W/m K)

cv_wall_IntMass::FT: Internal mass wall volumetric heat capacity (J/m³ K)

UrbanTethysChloris.ModelComponents.Parameters.WindowParameters — Type

WindowParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for building windows.

Fields

WindowsOn::Int: Include windows in simulation (0/1)

GlazingRatio::FT: Window-to-wall ratio (-)

Uvalue::FT: U-value of windows (W/m² K)

lan_windows::FT: Thermal conductivity of windows (W/m K)

cv_glass::FT: Volumetric heat capacity of glass (J/m³ K)

dztot::FT: Total thickness of glass layers (m)

SHGC::FT: Solar heat gain coefficient (-)

SolarTransmittance::FT: Solar radiation transmittance through windows (-)

SolarAbsorptivity::FT: Fraction of solar radiation absorbed by window (-)

SolarAlbedo::FT: Window albedo (-)

UrbanTethysChloris.ModelComponents.Parameters.HVACParameters — Type

HVACParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for HVAC system.

Fields

ACon::Int: Enable air conditioning (0/1)

Heatingon::Int: Enable heating (0/1)

TsetpointCooling::FT: Cooling setpoint temperature (K)

TsetpointHeating::FT: Heating setpoint temperature (K)

RHsetpointCooling::FT: Cooling setpoint relative humidity (%)

RHsetpointHeating::FT: Heating setpoint relative humidity (%)

ACH::FT: Air changes per hour (1/h)

COPAC::FT: Coefficient of performance for AC (-)

COPHeat::FT: Coefficient of performance for heating (-)

f_ACLatentToQ::FT: Fraction of latent heat removed by AC that is condensed (-)

UrbanTethysChloris.ModelComponents.Parameters.SoilParameters — Type

SoilParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Container for soil parameters for different urban surface components.

Fields

roof::VegetatedSoilParameters{FT}: Roof soil parameters

ground::VegetatedSoilParameters{FT}: Ground soil parameters

Sp_In_T::FT: Specific water retained by a tree (mm m^2 VEG area m^-2 plant area)

UrbanTethysChloris.ModelComponents.Parameters.VegetatedSoilParameters — Type

VegetatedSoilParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Soil parameters specific to vegetated surfaces (roof and ground)

Fields

Pcla::FT: Fraction of clay in the soil (-)

Psan::FT: Fraction of sand in the soil (-)

Porg::FT: Fraction of organic material in the soil (-)

In_max_imp::FT: Maxiumum interception capacity of impervious area (mm)

In_max_ground::FT: Maxiumum interception capacity of ground under roof vegetation (mm)

In_max_underveg::FT: Maxiumum interception capacity of vegetated ground area (mm)

In_max_bare::FT: Maxiumum interception capacity of bare ground area (mm)

Sp_In::FT: specific water retained by a vegetated surface (mm m^2 VEG area m^-2 plant area)

Kimp::FT: Hydraulic conductivity of impervious area (mm/h)

Kfc::FT: Conductivity at field capacity (mm/h)

Phy::FT: Suction at the residual/hygroscopic water content (kPa)

SPAR::Int: SOIL PARAMETER TYPE 1-VanGenuchten 2-Saxton-Rawls

Kbot::FT: Conductivity at the bedrock layer (mm/h)

UrbanTethysChloris.ModelComponents.Parameters.PersonParameters — Type

PersonParameters{FT<:AbstractFloat} <: AbstractParameters{FT}

Parameters for the person in the urban canyon for MRT calculations.

Fields

PositionPx::FT: Position within canyon [m]

PositionPz::FT: Height of centre of person [m]

PersonWidth::FT: Horizontal radius of ellipse describing person (=hip width / 2) [-]

PersonHeight::FT: Vertical radius of ellipse describing person (= height / 2) [-]

HeightWind::FT: Height for wind speed to calculate OTC [m]

UrbanTethysChloris.ModelComponents.Parameters.ParameterSet — Type

ParameterSet{FT<:AbstractFloat} <: AbstractParameterSet{FT}

Parameters for the Urban Tethys-Chloris model.

Fields

building_energy::BuildingEnergyModelParameters{FT}: Parameters for the building energy model.

person::PersonParameters{FT}: Parameters for the person.

soil::SoilParameters{FT}: Parameters for the soil.

surfacefractions::SurfaceFractions{FT}: Parameters for the surface fractions.

thermal::ThermalProperties{FT}: Parameters for the thermal properties.

optical::OpticalProperties{FT}: Parameters for the optical properties.

urbangeometry::UrbanGeometryParameters{FT}: Parameters for the urban geometry.

vegetation::VegetationParameters{FT}: Parameters for the vegetation.